reorganized the structure of functions
